Add cast to quiet -Wshorten-64-to-32 warning
Clears clang 13 warning of:
math.c:433:29: warning: implicit conversion loses integer precision:
 'unsigned long' to 'int' [-Wshorten-64-to-32]
    exponent=strlen(dispstr)-1;  /* where the '-' goes */
            ~~~~~~~~~~~~~~~~^~

Part-of: <https://gitlab.freedesktop.org/xorg/app/xcalc/-/merge_requests/15>

+++ math.c
@@ -430,7 +430,7 @@ eef(void)
   if (!exponent) {
     strcat(dispstr,"E+");
     DrawDisplay();
-    exponent=strlen(dispstr)-1;  /* where the '-' goes */
+    exponent=(int) strlen(dispstr)-1;  /* where the '-' goes */
   }
   clrdisp=0;
   entered=1;
